fix: fix sign overflow causing writing junk data to /sys/kernel/meminfo

+++ sys/kernel_subsystem.h
@@ -1,7 +1,6 @@
 #ifndef KERNEL_SUBSYSTEM_H
 #define KERNEL_SUBSYSTEM_H
 
-#include "types.h"
 #include <stdint.h>
 #include <stdbool.h>
 #include <stddef.h>
@@ -11,8 +10,8 @@
 
 typedef struct {
     char name[64];
-    ssize_t (*read)(char *buffer, size_t size, size_t offset);
-    ssize_t (*write)(const char *buffer, size_t size, size_t offset);
+    int (*read)(char *buffer, size_t size, size_t offset);
+    int (*write)(const char *buffer, size_t size, size_t offset);
 } subsystem_file_t;
 
 typedef struct {
@@ -23,8 +22,8 @@ typedef struct {
 
 void subsystem_register(const char *name, kernel_subsystem_t **out_sub);
 void subsystem_add_file(kernel_subsystem_t *sub, const char *name, 
-                        ssize_t (*read)(char*, size_t, size_t), 
-                        ssize_t (*write)(const char*, size_t, size_t));
+                        int (*read)(char*, size_t, size_t), 
+                        int (*write)(const char*, size_t, size_t));
 
 kernel_subsystem_t* subsystem_get_by_name(const char *name);
 int subsystem_get_count(void);
